Add 'restart_policy' to docker module
Since version 1.2.0/ApiVersion 1.14 docker supports the flag
`restart_policy`. docker.py supports that flag since version 0.5.0.

This adds the new module parameter, to take advantage of the progress
on dockers side.

I'm not overly happy with the argument_spec for this, since it should be
a choice, but it also needs the numbered parameter for on-failure (see:
https://github.com/docker/docker-py/blob/master/docs/api.md#start ).

I chose to ignore values other than the defined ones ('always', 'on-failure').
The outcome of this is, that docker will assume the "default"
dict(Name='', MaximumRetryCount=0) in such instances.